% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Brazil_death_anxiety.R
\docType{data}
\name{Brazil_death_anxiety}
\alias{Brazil_death_anxiety}
\title{Brazil death anxiety}
\format{
A data frame with 200 rows and 28 variables.
\describe{
  \item{participant}{Participant id}
  \item{gender}{Gender of pariticipant.}
  \item{year_born}{Year participant was born.}
  \item{age_in_2015}{Age of participant when data was collected.}
  \item{ethnicity}{Ethnicity	of participant.}
  \item{marital_status}{Marital status of participant.}
  \item{employment}{Employment status of participant.}
  \item{occupation}{Occupation of participant.}
  \item{fivefears_1}{The first entry in a free list of participants’ fears and worries.}
  \item{fivefears_2}{The second entry in a free list of participants’ fears and worries.}
  \item{fivefears_3}{The third entry in a free list of participants’ fears and worries.}
  \item{fivefears_4}{The fourth entry in a free list of participants’ fears and worries.}
  \item{fivefears_5}{The fifth entry in a free list of participants’ fears and worries.}
  \item{worry_death}{Do you worry about dying?}
  \item{incomplete_bucket_list}{Does it bother you that you may die before you have done everything you wanted to?}
  \item{deathly_ill}{Do you worry that you may be very ill for a long time before you die?}
  \item{concern_for_others}{Does it upset you to think that others may see you suffering when you die?}
  \item{painful_death}{Do you worry that dying may be very painful?}
  \item{loved_ones}{Do you worry that the persons most close to you won’t be with you when you are dying?}
  \item{die_alone}{Do you worry that you may be alone when you are dying?}
  \item{cognitive_function}{Does the thought bother you that you might lose control of your mind before death?}
  \item{expenses_burden}{Do you worry that expenses connected with your dying will be a burden for other people?}
  \item{as_you_wish}{Does it worry you that your instructions or will about your belongings may not be carried out after you die?}
  \item{buried_alive}{Are you afraid that you may be buried before you are really dead?}
  \item{left_behind}{Does the thought of leaving loved ones behind you when you die disturb you?}
  \item{remembered}{Do you worry that those you care about may not remember you after your death?}
  \item{permenant}{Does the thought worry you that with death you may be gone forever?}
  \item{after_death}{Are you worried about not knowing what to expect after death?}
}
}
\source{
Jong, J., Halberstadt, J., Bluemke, M. et al. Death anxiety, exposure to death, mortuary preferences, and religiosity in five countries. Sci Data 6, 154 (2019). [Link to doi](https://doi.org/10.1038/s41597-019-0163-x)
}
\usage{
Brazil_death_anxiety
}
\description{
The dataset consists of responses from 200 individuals from Brazil.
The main measure is the DAQ, a 15-item multidimensional scale that
measures fears related to suffering, loneliness, the unknown, and
personal extinction (see Variable Description). The scale measures
range from 0 (“not at all”) to 8 (“very much”). In addition, the
file contains a free-list measure of participants’ fears and worries,
which may be coded, and several demographic variables.
}
\keyword{datasets}
